Version 1.7.5
FossilOrigin-Name: f23add0d245f16124549f8875bf7b5c26b94fa82a44b3c605747dd4839f7f583

+++ examples/custom1.cc
@@ -8,7 +8,7 @@
 sql_create_5 (stock,		// struct name, 
 	      1, 5,		// I'll explain these latter
 	      string, item,	// type, id
-	      int, num,
+	      longlong, num,
 	      double, weight,
 	      double, price,
 	      Date, sdate)
